Update materials repo to have extra code example.
